requirements_raspberry.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorclass==2.2.2
cryptography==44.0.0
easygui==0.98.3
flatbuffers==25.1.21
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib-metadata==8.5.0
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
msoffcrypto-tool==5.4.2
numpy==1.24.3
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==24.2
pcodedmp==1.2.6
protobuf==4.25.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-lite==2.13.0  # Usar TensorFlow Lite en lugar de TensorFlow completo
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.5.0
urllib3==2.2.3
werkzeug==3.0.6
win-unicode-console==0.5
wrapt==1.17.2
zipp==3.20.2
